+++ skdue_calendar/models/__init__.py
@@ -1,0 +1,2 @@
+from .calendar import Calendar
+from .calendar_event import CalendarEvent
