+++ skdue_calendar/serializers/__init__.py
@@ -1,0 +1,2 @@
+from .calendar_serializer import CalendarSerializer
+from .calendar_event_serializer import CalendarEventSerializer
