+++ skdue_calendar/serializers/calendar_event_serializer.py
@@ -1,5 +1,5 @@
 from rest_framework import serializers
-from .models import Calendar, CalendarEvent
+from skdue_calendar.models import CalendarEvent
 
 
 class CalendarEventSerializer(serializers.ModelSerializer):
@@ -13,15 +13,4 @@ class CalendarEventSerializer(serializers.ModelSerializer):
             "description",
             "start_date",
             "end_date"
-        )
-
-
-class CalendarSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Calendar
-        fields = (
-            "id",
-            "name",
-            "slug",
-            "get_absolute_url"
         )
