+++ skdue_calendar/serializers/calendar_serializer.py
@@ -1,0 +1,13 @@
+from rest_framework import serializers
+from skdue_calendar.models import Calendar
+
+
+class CalendarSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Calendar
+        fields = (
+            "id",
+            "name",
+            "slug",
+            "get_absolute_url"
+        )
