+++ skdue_calendar/views/__init__.py
@@ -1,0 +1,4 @@
+from .calendar_list import CalendarList
+from .calendar_event_list import CalendarEventList
+from .calendar_event_detail import CalendarEventDetail
+from .search import Search
